refactor locations and namespaces
