+++ crypto-scraper/app/Console/Base/Bitcointalk/BitcointalkParser.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Console;
+namespace App\Console\Base\Bitcointalk;
 
+use App\Console\Base\CryptoParser;
 use Symfony\Component\DomCrawler\Crawler;
 
 class BitcointalkParser extends CryptoParser {
