+++ crypto-scraper/app/Console/Base/Bitcointalk/KafkaConProducer.php
@@ -1,9 +1,8 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Console\Commands\Bitcointalk;
+namespace App\Console\Base\Bitcointalk;
 
-use App\Console\BitcointalkParser;
 use App\Kafka\ConProducerFeatures;
 
 abstract class KafkaConProducer extends BitcointalkParser {
