+++ crypto-scraper/app/Console/Commands/Bitcoinabuse/Load.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Console\Commands\Bitcoinabuse;
 
-use App\Console\CryptoParser;
+use App\Console\Base\CryptoParser;
 
 class Load extends CryptoParser {
     protected $signature = 'bitcoinabuse:load {verbose=1}';
