+++ crypto-scraper/app/Console/Commands/Bitcointalk/Parsers/TopicMessages.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 namespace App\Console\Commands\Bitcointalk\Parsers;
 
 use App\AddressMatcher;
-use App\Console\BitcointalkParser;
+use App\Console\Base\Bitcointalk\BitcointalkParser;
 use App\Console\Commands\Bitcointalk\UrlValidations;
-use App\Console\CryptoCurrency;
+use App\Console\Base\CryptoCurrency;
 use App\Models\ParsedAddress;
 use App\Models\Pg\Category;
 use Illuminate\Support\Arr;
