+++ crypto-scraper/app/Console/Commands/Bitcointalk/Runners/Boards.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Console\Commands\Bitcointalk\Runners;
 
-use App\Console\BitcointalkParser;
+use App\Console\Base\Bitcointalk\BitcointalkParser;
 use App\Models\Pg\Bitcointalk\BoardPage;
 
 class Boards extends BitcointalkParser {
