+++ crypto-scraper/app/Console/Commands/WriteLocalStorage.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace App\Console\Commands;
 
 use Illuminate\Support\Facades\Storage;
-use App\Console\CryptoParser;
+use App\Console\Base\CryptoParser;
 
 class WriteLocalStorage extends CryptoParser
 {
